+++ src/widgets/splits/InputCompletionPopup.hpp
@@ -8,11 +8,30 @@
 
 #include <functional>
 #include <memory>
+#include <vector>
 
 namespace chatterino {
 
 class Channel;
 using ChannelPtr = std::shared_ptr<Channel>;
+
+struct Emote;
+using EmotePtr = std::shared_ptr<const Emote>;
+
+namespace detail {
+
+    struct CompletionEmote {
+        EmotePtr emote;
+        QString displayName;
+        QString providerName;
+    };
+
+    std::vector<CompletionEmote> buildCompletionEmoteList(const QString &text,
+                                                          ChannelPtr channel);
+
+}  // namespace detail
+
+class GenericListView;
 
 class InputCompletionPopup : public BasePopup
 {