+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## Unversioned
 
+- Minor: Added `/shoutout <username>` commands to shoutout specified user. (#4638)
 - Dev: Added command to set Qt's logging filter/rules at runtime (`/c2-set-logging-rules`). (#4637)
 - Dev: Added the ability to see & load custom themes from the Themes directory. No stable promises are made of this feature, changes might be made that breaks custom themes without notice. (#4570)
+- Dev: Added test cases for emote and tab completion. (#4644)
+- Dev: Fixed `clang-tidy-review` action not picking up dependencies. (#4648)
 
 ## 2.4.4